SmartChatAI/ai-service/requirements.txt
Flask==2.3.2
scikit-learn==1.2.2
nltk==3.8.1
numpy==1.24.3
joblib==1.2.0
python-dotenv==1.0.0
requests==2.31.0
